Avoid printf error due to size of semaphore key type
this type is not necessarily any particular width, so this is the
safe way to get it printed.

+++ src/libnml/os_intf/_sem.c
@@ -20,6 +20,7 @@
 #include <stdarg.h>		/* va_list, va_arg(), va_start(), va_end() */
 #include <sys/types.h>
 #include <sys/ipc.h>		/* IPC_CREATE, IPC_NOWAIT */
+#include <inttypes.h>
 
 /* There are two types of posix semaphores named and unnamed.
    unamed semaphores can either have the pshared flag set or not
@@ -112,7 +113,7 @@ rcs_sem_t *rcs_sem_open(key_t name, int oflag, /* int mode */ ...)
     key = name;
 
     if (key < 1) {
-	rcs_print_error("rcs_sem_open: invalid key %d\n", key);
+	rcs_print_error("rcs_sem_open: invalid key %jd\n", (intmax_t)key);
 	return NULL;
     }
